+++ test/object-store/sync/flx_sync.cpp
@@ -45,19 +45,6 @@
 #include <iostream>
 #include <stdexcept>
 
-namespace realm {
-
-class TestHelper {
-public:
-    static bool can_advance(const SharedRealm& realm)
-    {
-        auto& coord = Realm::Internal::get_coordinator(*realm);
-        return coord.can_advance(*realm);
-    }
-};
-
-} // namespace realm
-
 namespace realm::app {
 
 namespace {
@@ -117,37 +104,25 @@ std::vector<ObjectId> fill_large_array_schema(FLXSyncTestHarness& harness)
     return ret;
 }
 
-void wait_for_advance(const SharedRealm& realm)
+void wait_for_error_to_persist(const AppSession& app_session, const std::string& err)
 {
-    timed_wait_for([&] {
-        return !TestHelper::can_advance(realm);
-    });
-}
-
 // TODO: Re-enable it in RCORE-1241.
 #if 0
-// Returns true if `err` is among the erros received from the server, false otherwise.
-bool wait_for_error_to_persist(const AppSession& app_session, const std::string& err)
-{
-    bool error_found = false;
     timed_sleeping_wait_for(
         [&]() -> bool {
             auto errors = app_session.admin_api.get_errors(app_session.server_app_id);
-            auto it = std::find_if(errors.begin(), errors.end(), [&err](const auto& error) {
-                return err == error;
-            });
-
+            auto it = std::find(errors.begin(), errors.end(), err);
             if (it == errors.end()) {
                 millisleep(500); // don't spam the server too much
             }
-            error_found = it != errors.end();
-            return error_found;
+            return it != errors.end();
         },
         std::chrono::minutes(10));
-
-    return error_found;
-}
+#else
+    static_cast<void>(app_session);
+    static_cast<void>(err);
 #endif
+}
 } // namespace
 
 TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
@@ -192,7 +167,7 @@ TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
 
         wait_for_download(*realm);
         {
-            wait_for_advance(realm);
+            realm->refresh();
             Results results(realm, table);
             CHECK(results.size() == 1);
             auto obj = results.get<Obj>(0);
@@ -210,7 +185,7 @@ TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
         }
 
         {
-            wait_for_advance(realm);
+            realm->refresh();
             Results results(realm, Query(table));
             CHECK(results.size() == 2);
         }
@@ -228,7 +203,7 @@ TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
         }
 
         {
-            wait_for_advance(realm);
+            realm->refresh();
             Results results(realm, Query(table));
             CHECK(results.size() == 1);
             auto obj = results.get<Obj>(0);
@@ -244,7 +219,7 @@ TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
         }
 
         {
-            wait_for_advance(realm);
+            realm->refresh();
             Results results(realm, table);
             CHECK(results.size() == 0);
         }
@@ -857,7 +832,7 @@ TEST_CASE("flx: uploading an object that is out-of-view results in compensating 
                 std::move(error_future).get(), invalid_obj,
                 util::format("write to '%1' in table \"TopLevel\" not allowed", invalid_obj.to_string()));
 
-            wait_for_advance(realm);
+            realm->refresh();
 
             auto top_level_table = realm->read_group().get_table("class_TopLevel");
             REQUIRE(top_level_table->is_empty());
@@ -901,7 +876,7 @@ TEST_CASE("flx: uploading an object that is out-of-view results in compensating 
                 std::move(error_future).get(), invalid_obj,
                 util::format("write to '%1' in table \"TopLevel\" not allowed", invalid_obj.to_string()));
 
-            wait_for_advance(realm);
+            realm->refresh();
 
             obj = Object::get_for_primary_key(c, realm, "TopLevel", std::any(invalid_obj));
             embedded_obj = util::any_cast<Object&&>(obj.get_property_value<std::any>(c, "embedded_obj"));
@@ -917,7 +892,7 @@ TEST_CASE("flx: uploading an object that is out-of-view results in compensating 
             wait_for_upload(*realm);
             wait_for_download(*realm);
 
-            wait_for_advance(realm);
+            realm->refresh();
             obj = Object::get_for_primary_key(c, realm, "TopLevel", std::any(invalid_obj));
             embedded_obj = util::any_cast<Object&&>(obj.get_property_value<std::any>(c, "embedded_obj"));
             REQUIRE(util::any_cast<std::string&&>(embedded_obj.get_property_value<std::any>(c, "str_field")) ==
@@ -960,7 +935,7 @@ TEST_CASE("flx: uploading an object that is out-of-view results in compensating 
             validate_sync_error(std::move(error_future).get(), invalid_obj,
                                 "object is outside of the current query view");
 
-            wait_for_advance(realm);
+            realm->refresh();
 
             auto top_level_table = realm->read_group().get_table("class_TopLevel");
             REQUIRE(top_level_table->size() == 1);
@@ -1078,7 +1053,7 @@ TEST_CASE("flx: interrupted bootstrap restarts/recovers on reconnect", "[sync][f
     wait_for_upload(*realm);
     wait_for_download(*realm);
 
-    wait_for_advance(realm);
+    realm->refresh();
     REQUIRE(table->size() == obj_ids_at_end.size());
     for (auto& id : obj_ids_at_end) {
         REQUIRE(table->find_primary_key(Mixed{id}));
@@ -1476,7 +1451,7 @@ TEST_CASE("flx: bootstrap batching prevents orphan documents", "[sync][flx][app]
     auto mutate_realm = [&] {
         harness.load_initial_data([&](SharedRealm realm) {
             auto table = realm->read_group().get_table("class_TopLevel");
-            wait_for_advance(realm);
+            realm->refresh();
             Results res(realm, Query(table).greater(table->get_column_key("queryable_int_field"), int64_t(10)));
             REQUIRE(res.size() == 2);
             res.clear();
@@ -1614,7 +1589,7 @@ TEST_CASE("flx: bootstrap batching prevents orphan documents", "[sync][flx][app]
         wait_for_upload(*realm);
         wait_for_download(*realm);
 
-        wait_for_advance(realm);
+        realm->refresh();
         auto expected_obj_ids = util::Span<ObjectId>(obj_ids_at_end).sub_span(0, 3);
 
         REQUIRE(table->size() == expected_obj_ids.size());
@@ -1726,7 +1701,7 @@ TEST_CASE("flx: bootstrap batching prevents orphan documents", "[sync][flx][app]
             .get();
         wait_for_upload(*realm);
         wait_for_download(*realm);
-        wait_for_advance(realm);
+        realm->refresh();
 
         auto expected_obj_ids = util::Span<ObjectId>(obj_ids_at_end).sub_span(0, 3);
 
@@ -1894,13 +1869,10 @@ TEST_CASE("flx: asymmetric sync", "[sync][flx][app]") {
             CHECK(ec.value() == int(realm::sync::ClientError::bad_changeset));
         });
 
-// TODO: Re-enable it in RCORE-1241.
-#if 0
-        REQUIRE(wait_for_error_to_persist(
+        REQUIRE_NOTHROW(wait_for_error_to_persist(
             harness->session().app_session(),
             "Failed to transform received changeset: Schema mismatch: 'Asymmetric' is asymmetric "
             "on one side, but not on the other. (ProtocolErrorCode=112)"));
-#endif
     }
 
     SECTION("basic embedded object construction") {
@@ -2002,8 +1974,6 @@ TEST_CASE("flx: asymmetric sync - dev mode", "[sync][flx][app]") {
 }
 #endif
 
-// TODO: Re-enable it in RCORE-1241.
-#if 0
 TEST_CASE("flx: send client error", "[sync][flx][app]") {
     FLXSyncTestHarness harness("flx_client_error");
 
@@ -2021,9 +1991,9 @@ TEST_CASE("flx: send client error", "[sync][flx][app]") {
 
     error_future.get();
 
-    REQUIRE(wait_for_error_to_persist(harness.session().app_session(), "simulated failure (ProtocolErrorCode=112)"));
+    REQUIRE_NOTHROW(
+        wait_for_error_to_persist(harness.session().app_session(), "simulated failure (ProtocolErrorCode=112)"));
 }
-#endif
 
 } // namespace realm::app